requirement.txt
python=3.8
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apturl==0.5.2
asttokens==2.2.1
backcall==0.2.0
black==21.4b2
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
# Editable install with no version control (detectron2==0.6)
-e /home/cz/detectron2
distro==1.4.0
distro-info===0.23ubuntu1
easyocr==1.7.0
entrypoints==0.3
executing==1.2.0
filelock==3.9.0
fonttools==4.40.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
httplib2==0.14.0
hydra-core==1.3.2
idna==2.8
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
kafka-python3==3.0.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.3
lit==15.0.7
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
mypy-extensions==1.0.0
netifaces==0.10.4
networkx==3.0
ninja==1.11.1
numpy==1.23.5
oauthlib==3.1.0
omegaconf==2.3.0
opencv-contrib-python @ file:///home/cz/project/inityolov7/opencv_contrib_python-4.8.0.76-cp37-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.21.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.5.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycups==1.9.73
pydot==1.4.2
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
stack-data==0.6.2
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tifffile==2023.4.12
toml==0.10.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.6
Werkzeug==2.3.6
xkit==0.0.0
yacs==0.1.8
zipp==3.15.0
